Remove css catcher
